move mobile nav specific functions to mobile nav file
